[854] Fixes unexplained failure with fetching publisher information.
Whilst we are still using the google lib we only let it do the oauth dance
before we take over making the requests using the python-requests library.

In cases where it fails with a 401 we will retry the request ONCE after
trying to obtain a new token.

Re-enabled all of the relevant links and html blocks that were commented out.